Actualización después de pruebas en postman, de endpoints para trato de usuarios

+++ src/main/java/com/jve/Service/UsuarioService.java
@@ -52,7 +52,11 @@ public class UsuarioService {
         });
     }
 
-    public void deleteUsuario(Integer id) {
-        usuarioRepository.deleteById(id);
+    public boolean deleteUsuario(Integer id) {
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
